+++ CHANGELOG.md
@@ -11,9 +11,11 @@ Version 1.1.0-SNAPSHOT
 - Installation script for Tesseract OCR on Windows using Batch (without Chocolatey)
 - Installation script for Tesseract OCR on Windows with Chocolatey (via Bash)
 - New field in `ApplicationModelBean` for the selected PDF file
+- new dependency `io.github.astrapi69:jobj-core` in version 9.1
 
 ### Changed
 - Updated Lombok dependency to patch version `1.18.36`
+- Upgraded Gradle plugin `io.freefair.lombok` to version 8.11
 - Updated dependency `pdf-to-text-extensions` to minor version `1.4`
 - Updated dependency `silly-io` to minor version `3.4`